Refactoring suppression examples
