+++ examples/suppression/GetAllSpamReports.java
@@ -1,0 +1,33 @@
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+import com.sendgrid.*;
+
+import java.io.IOException;
+import java.util.HashMap;
+import java.util.Map;
+
+//////////////////////////////////////////////////////////////////
+// Retrieve all spam reports
+// GET /suppression/spam_reports
+
+public class GetAllSpamReports {
+    public static void main(String[] args) throws IOException {
+        try {
+            SendGrid sg = new SendGrid(System.getenv("SENDGRID_API_KEY"));
+            Request request = new Request();
+            request.setMethod(Method.GET);
+            request.setEndpoint("suppression/spam_reports");
+            request.addQueryParam("start_time", "1");
+            request.addQueryParam("limit", "1");
+            request.addQueryParam("end_time", "1");
+            request.addQueryParam("offset", "1");
+            Response response = sg.api(request);
+            System.out.println(response.getStatusCode());
+            System.out.println(response.getBody());
+            System.out.println(response.getHeaders());
+        } catch (IOException ex) {
+            throw ex;
+        }
+    }
+}
