+++ examples/suppression/README.md
@@ -1,0 +1,21 @@
+![SendGrid Logo](https://uiux.s3.amazonaws.com/2016-logos/email-logo%402x.png)
+
+This folder contains various examples on using the SUPPRESSION endpoint of SendGrid with Java:
+
+* [Retrieve all blocks (GET /suppression/blocks)](GetAllBlocks.java)
+* [Delete blocks (DELETE /suppression/blocks)](DeleteBlocks.java)
+* [Retrieve a specific block (GET /suppression/blocks/{email})](GetSpecificBlock.java)
+* [Delete a specific block (DELETE /suppression/blocks/{email})](DeleteSpecificBlock.java)
+* [Retrieve all bounces (GET /suppression/bounces)](GetAllBounces.java)
+* [Delete bounces (DELETE /suppression/bounces)](DeleteBounces.java)
+* [Retrieve a Bounce (GET /suppression/bounces/{email})](GetBounce.java)
+* [Delete a bounce (DELETE /suppression/bounces/{email})](DeleteBounce.java)
+* [Retrieve all invalid emails (GET /suppression/invalid_emails)](GetAllInvalidEmails.java)
+* [Delete invalid emails (DELETE /suppression/invalid_emails)](DeleteInvalidEmails.java)
+* [Retrieve a specific invalid email (GET /suppression/invalid_emails/{email})](GetSpecificInvalidEmail.java)
+* [Delete a specific invalid email (DELETE /suppression/invalid_emails/{email})](DeleteSpecificInvalidEmail.java)
+* [Retrieve a specific spam report (GET /suppression/spam_report/{email})](GetSpecificSpamReport.java)
+* [Delete a specific spam report (DELETE /suppression/spam_report/{email})](DeleteSpecificSpamReport.java)
+* [Retrieve all spam reports (GET /suppression/spam_reports)](GetAllSpamReports.java)
+* [Delete spam reports (DELETE /suppression/spam_reports)](DeleteSpamReports.java)
+* [Retrieve all global suppressions (GET /suppression/unsubscribes)](GetAllGlobalSuppressions.java)
