--- examples/suppression/suppression.java
@@ -1,406 +1,0 @@
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-
-import com.sendgrid.*;
-
-import java.io.IOException;
-import java.util.HashMap;
-import java.util.Map;
-
-//////////////////////////////////////////////////////////////////
-// Retrieve all blocks
-// GET /suppression/blocks
-
-
-public class Example {
-  public static void main(String[] args) throws IOException {
-    try {
-      SendGrid sg = new SendGrid(System.getenv("SENDGRID_API_KEY"));
-      Request request = new Request();
-      request.setMethod(Method.GET);
-      request.setEndpoint("suppression/blocks");
-      request.addQueryParam("start_time", "1");
-      request.addQueryParam("limit", "1");
-      request.addQueryParam("end_time", "1");
-      request.addQueryParam("offset", "1");
-      Response response = sg.api(request);
-      System.out.println(response.getStatusCode());
-      System.out.println(response.getBody());
-      System.out.println(response.getHeaders());
-    } catch (IOException ex) {
-      throw ex;
-    }
-  }
-}
-
-//////////////////////////////////////////////////////////////////
-// Delete blocks
-// DELETE /suppression/blocks
-
-
-public class Example {
-  public static void main(String[] args) throws IOException {
-    try {
-      SendGrid sg = new SendGrid(System.getenv("SENDGRID_API_KEY"));
-      Request request = new Request();
-      request.setMethod(Method.DELETE);
-      request.setEndpoint("suppression/blocks");
-      request.setBody("{\"emails\":[\"example1@example.com\",\"example2@example.com\"],\"delete_all\":false}");
-      Response response = sg.api(request);
-      System.out.println(response.getStatusCode());
-      System.out.println(response.getBody());
-      System.out.println(response.getHeaders());
-    } catch (IOException ex) {
-      throw ex;
-    }
-  }
-}
-
-//////////////////////////////////////////////////////////////////
-// Retrieve a specific block
-// GET /suppression/blocks/{email}
-
-
-public class Example {
-  public static void main(String[] args) throws IOException {
-    try {
-      SendGrid sg = new SendGrid(System.getenv("SENDGRID_API_KEY"));
-      Request request = new Request();
-      request.setMethod(Method.GET);
-      request.setEndpoint("suppression/blocks/{email}");
-      Response response = sg.api(request);
-      System.out.println(response.getStatusCode());
-      System.out.println(response.getBody());
-      System.out.println(response.getHeaders());
-    } catch (IOException ex) {
-      throw ex;
-    }
-  }
-}
-
-//////////////////////////////////////////////////////////////////
-// Delete a specific block
-// DELETE /suppression/blocks/{email}
-
-
-public class Example {
-  public static void main(String[] args) throws IOException {
-    try {
-      SendGrid sg = new SendGrid(System.getenv("SENDGRID_API_KEY"));
-      Request request = new Request();
-      request.setMethod(Method.DELETE);
-      request.setEndpoint("suppression/blocks/{email}");
-      Response response = sg.api(request);
-      System.out.println(response.getStatusCode());
-      System.out.println(response.getBody());
-      System.out.println(response.getHeaders());
-    } catch (IOException ex) {
-      throw ex;
-    }
-  }
-}
-
-//////////////////////////////////////////////////////////////////
-// Retrieve all bounces
-// GET /suppression/bounces
-
-
-public class Example {
-  public static void main(String[] args) throws IOException {
-    try {
-      SendGrid sg = new SendGrid(System.getenv("SENDGRID_API_KEY"));
-      Request request = new Request();
-      request.setMethod(Method.GET);
-      request.setEndpoint("suppression/bounces");
-      request.addQueryParam("start_time", "1");
-      request.addQueryParam("end_time", "1");
-      Response response = sg.api(request);
-      System.out.println(response.getStatusCode());
-      System.out.println(response.getBody());
-      System.out.println(response.getHeaders());
-    } catch (IOException ex) {
-      throw ex;
-    }
-  }
-}
-
-//////////////////////////////////////////////////////////////////
-// Delete bounces
-// DELETE /suppression/bounces
-
-
-public class Example {
-  public static void main(String[] args) throws IOException {
-    try {
-      SendGrid sg = new SendGrid(System.getenv("SENDGRID_API_KEY"));
-      Request request = new Request();
-      request.setMethod(Method.DELETE);
-      request.setEndpoint("suppression/bounces");
-      request.setBody("{\"emails\":[\"example@example.com\",\"example2@example.com\"],\"delete_all\":true}");
-      Response response = sg.api(request);
-      System.out.println(response.getStatusCode());
-      System.out.println(response.getBody());
-      System.out.println(response.getHeaders());
-    } catch (IOException ex) {
-      throw ex;
-    }
-  }
-}
-
-//////////////////////////////////////////////////////////////////
-// Retrieve a Bounce
-// GET /suppression/bounces/{email}
-
-
-public class Example {
-  public static void main(String[] args) throws IOException {
-    try {
-      SendGrid sg = new SendGrid(System.getenv("SENDGRID_API_KEY"));
-      Request request = new Request();
-      request.setMethod(Method.GET);
-      request.setEndpoint("suppression/bounces/{email}");
-      Response response = sg.api(request);
-      System.out.println(response.getStatusCode());
-      System.out.println(response.getBody());
-      System.out.println(response.getHeaders());
-    } catch (IOException ex) {
-      throw ex;
-    }
-  }
-}
-
-//////////////////////////////////////////////////////////////////
-// Delete a bounce
-// DELETE /suppression/bounces/{email}
-
-
-public class Example {
-  public static void main(String[] args) throws IOException {
-    try {
-      SendGrid sg = new SendGrid(System.getenv("SENDGRID_API_KEY"));
-      Request request = new Request();
-      request.setMethod(Method.DELETE);
-      request.setEndpoint("suppression/bounces/{email}");
-      request.addQueryParam("email_address", "example@example.com");
-      Response response = sg.api(request);
-      System.out.println(response.getStatusCode());
-      System.out.println(response.getBody());
-      System.out.println(response.getHeaders());
-    } catch (IOException ex) {
-      throw ex;
-    }
-  }
-}
-
-//////////////////////////////////////////////////////////////////
-// Retrieve all invalid emails
-// GET /suppression/invalid_emails
-
-
-public class Example {
-  public static void main(String[] args) throws IOException {
-    try {
-      SendGrid sg = new SendGrid(System.getenv("SENDGRID_API_KEY"));
-      Request request = new Request();
-      request.setMethod(Method.GET);
-      request.setEndpoint("suppression/invalid_emails");
-      request.addQueryParam("start_time", "1");
-      request.addQueryParam("limit", "1");
-      request.addQueryParam("end_time", "1");
-      request.addQueryParam("offset", "1");
-      Response response = sg.api(request);
-      System.out.println(response.getStatusCode());
-      System.out.println(response.getBody());
-      System.out.println(response.getHeaders());
-    } catch (IOException ex) {
-      throw ex;
-    }
-  }
-}
-
-//////////////////////////////////////////////////////////////////
-// Delete invalid emails
-// DELETE /suppression/invalid_emails
-
-
-public class Example {
-  public static void main(String[] args) throws IOException {
-    try {
-      SendGrid sg = new SendGrid(System.getenv("SENDGRID_API_KEY"));
-      Request request = new Request();
-      request.setMethod(Method.DELETE);
-      request.setEndpoint("suppression/invalid_emails");
-      request.setBody("{\"emails\":[\"example1@example.com\",\"example2@example.com\"],\"delete_all\":false}");
-      Response response = sg.api(request);
-      System.out.println(response.getStatusCode());
-      System.out.println(response.getBody());
-      System.out.println(response.getHeaders());
-    } catch (IOException ex) {
-      throw ex;
-    }
-  }
-}
-
-//////////////////////////////////////////////////////////////////
-// Retrieve a specific invalid email
-// GET /suppression/invalid_emails/{email}
-
-
-public class Example {
-  public static void main(String[] args) throws IOException {
-    try {
-      SendGrid sg = new SendGrid(System.getenv("SENDGRID_API_KEY"));
-      Request request = new Request();
-      request.setMethod(Method.GET);
-      request.setEndpoint("suppression/invalid_emails/{email}");
-      Response response = sg.api(request);
-      System.out.println(response.getStatusCode());
-      System.out.println(response.getBody());
-      System.out.println(response.getHeaders());
-    } catch (IOException ex) {
-      throw ex;
-    }
-  }
-}
-
-//////////////////////////////////////////////////////////////////
-// Delete a specific invalid email
-// DELETE /suppression/invalid_emails/{email}
-
-
-public class Example {
-  public static void main(String[] args) throws IOException {
-    try {
-      SendGrid sg = new SendGrid(System.getenv("SENDGRID_API_KEY"));
-      Request request = new Request();
-      request.setMethod(Method.DELETE);
-      request.setEndpoint("suppression/invalid_emails/{email}");
-      Response response = sg.api(request);
-      System.out.println(response.getStatusCode());
-      System.out.println(response.getBody());
-      System.out.println(response.getHeaders());
-    } catch (IOException ex) {
-      throw ex;
-    }
-  }
-}
-
-//////////////////////////////////////////////////////////////////
-// Retrieve a specific spam report
-// GET /suppression/spam_report/{email}
-
-
-public class Example {
-  public static void main(String[] args) throws IOException {
-    try {
-      SendGrid sg = new SendGrid(System.getenv("SENDGRID_API_KEY"));
-      Request request = new Request();
-      request.setMethod(Method.GET);
-      request.setEndpoint("suppression/spam_report/{email}");
-      Response response = sg.api(request);
-      System.out.println(response.getStatusCode());
-      System.out.println(response.getBody());
-      System.out.println(response.getHeaders());
-    } catch (IOException ex) {
-      throw ex;
-    }
-  }
-}
-
-//////////////////////////////////////////////////////////////////
-// Delete a specific spam report
-// DELETE /suppression/spam_report/{email}
-
-
-public class Example {
-  public static void main(String[] args) throws IOException {
-    try {
-      SendGrid sg = new SendGrid(System.getenv("SENDGRID_API_KEY"));
-      Request request = new Request();
-      request.setMethod(Method.DELETE);
-      request.setEndpoint("suppression/spam_report/{email}");
-      Response response = sg.api(request);
-      System.out.println(response.getStatusCode());
-      System.out.println(response.getBody());
-      System.out.println(response.getHeaders());
-    } catch (IOException ex) {
-      throw ex;
-    }
-  }
-}
-
-//////////////////////////////////////////////////////////////////
-// Retrieve all spam reports
-// GET /suppression/spam_reports
-
-
-public class Example {
-  public static void main(String[] args) throws IOException {
-    try {
-      SendGrid sg = new SendGrid(System.getenv("SENDGRID_API_KEY"));
-      Request request = new Request();
-      request.setMethod(Method.GET);
-      request.setEndpoint("suppression/spam_reports");
-      request.addQueryParam("start_time", "1");
-      request.addQueryParam("limit", "1");
-      request.addQueryParam("end_time", "1");
-      request.addQueryParam("offset", "1");
-      Response response = sg.api(request);
-      System.out.println(response.getStatusCode());
-      System.out.println(response.getBody());
-      System.out.println(response.getHeaders());
-    } catch (IOException ex) {
-      throw ex;
-    }
-  }
-}
-
-//////////////////////////////////////////////////////////////////
-// Delete spam reports
-// DELETE /suppression/spam_reports
-
-
-public class Example {
-  public static void main(String[] args) throws IOException {
-    try {
-      SendGrid sg = new SendGrid(System.getenv("SENDGRID_API_KEY"));
-      Request request = new Request();
-      request.setMethod(Method.DELETE);
-      request.setEndpoint("suppression/spam_reports");
-      request.setBody("{\"emails\":[\"example1@example.com\",\"example2@example.com\"],\"delete_all\":false}");
-      Response response = sg.api(request);
-      System.out.println(response.getStatusCode());
-      System.out.println(response.getBody());
-      System.out.println(response.getHeaders());
-    } catch (IOException ex) {
-      throw ex;
-    }
-  }
-}
-
-//////////////////////////////////////////////////////////////////
-// Retrieve all global suppressions
-// GET /suppression/unsubscribes
-
-
-public class Example {
-  public static void main(String[] args) throws IOException {
-    try {
-      SendGrid sg = new SendGrid(System.getenv("SENDGRID_API_KEY"));
-      Request request = new Request();
-      request.setMethod(Method.GET);
-      request.setEndpoint("suppression/unsubscribes");
-      request.addQueryParam("start_time", "1");
-      request.addQueryParam("limit", "1");
-      request.addQueryParam("end_time", "1");
-      request.addQueryParam("offset", "1");
-      Response response = sg.api(request);
-      System.out.println(response.getStatusCode());
-      System.out.println(response.getBody());
-      System.out.println(response.getHeaders());
-    } catch (IOException ex) {
-      throw ex;
-    }
-  }
-}
-
